servicio-2/requirements.txt
Flask
APScheduler 
loging
paho-mqtt
influxdb-client